requirements.txt
couchbase==4.3.0
datasets==3.0.0
langchain==0.3.0
langchain_core==0.3.5
langchain_couchbase==0.1.1
langchain_openai==0.2.0
python-dotenv==1.0.1
ragas==0.1.20
tqdm==4.66.4
